Fixed style

+++ tasks/rational/rational.cpp
@@ -56,16 +56,16 @@ void Rational::SetDenominator(int value) {
 }
 
 Rational& operator+=(Rational& lhs, const Rational& rhs) {
-    int64_t new_numer = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetNumerator()) +
-                        static_cast<int64_t>(rhs.GetDenominator()) * static_cast<int64_t>(lhs.GetNumerator());
-    int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetDenominator());
+    int64_t new_numer = static_cast<int64_t>(lhs.GetDenominator()) * rhs.GetNumerator() +
+                        static_cast<int64_t>(rhs.GetDenominator()) * lhs.GetNumerator();
+    int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * rhs.GetDenominator();
     lhs.Set(new_numer, new_denom);
     return lhs;
 }
 
 Rational& operator*=(Rational& lhs, const Rational& rhs) {
-    int64_t new_numer = static_cast<int64_t>(lhs.GetNumerator()) * static_cast<int64_t>(rhs.GetNumerator());
-    int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetDenominator());
+    int64_t new_numer = static_cast<int64_t>(lhs.GetNumerator()) * rhs.GetNumerator();
+    int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * rhs.GetDenominator();
     lhs.Set(new_numer, new_denom);
     return lhs;
 }
